Extract builder sections into separate concern modules
Split Builder::Spec into focused concerns:
- InfoBuilder: API info and license configuration
- ServerBuilder: Server URLs from host/basePath/schemes
- SecurityBuilder: Security schemes and OAuth flows
- OperationBuilder: Operation details (summary, produces, consumes)
- TagBuilder: Tag collection and merging

This improves maintainability by organizing related methods together.
